Added a semantic analysis test for including library .bond files. Fixed some bugs: - Symbols were added to the wrong scope when there are multiple occurences of a namespace with the same name. - Infinite loop in the type evaluation pass if some arguments fail to resolve. - Pretty printer only printed the fist element of qualified identifiers.

+++ source/libmath.cpp
@@ -128,7 +128,7 @@ void Atan2f(Bond::VM &vm)
 	VM::CalleeStackFrame &frame = vm.GetTopStackFrame();
 	const double y = double(frame.GetArg<bf32_t>(0));
 	const double x = double(frame.GetArg<bf32_t>(1));
-	const double result = atan2f(y, x);
+	const double result = atan2(y, x);
 	frame.SetReturnValue(bf32_t(result));
 }
